requirements.txt
# Core ML Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Gradient Boosting Models
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Chemical Informatics
rdkit>=2022.03.1  # Install with conda: conda install -c conda-forge rdkit

# Hyperparameter Optimization
optuna>=3.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
joblib>=1.1.0

# Web Framework (optional, for API)
fastapi>=0.85.0
uvicorn>=0.18.0

# Development Tools (optional)
jupyter>=1.0.0
ipython>=7.0.0
black>=22.0.0
flake8>=4.0.0
pytest>=7.0.0